+++ stregsystem/mail.py
@@ -1,6 +1,5 @@
 import smtplib
 import logging
-import csv
 
 
 from email.mime.multipart import MIMEMultipart
@@ -11,6 +10,7 @@ from django.template.loader import render_to_string
 from django.utils.html import escape
 from django.utils import timezone
 from stregsystem.templatetags.stregsystem_extras import money
+from stregsystem.utils import rows_to_csv
 
 logger = logging.getLogger(__name__)
 
@@ -34,21 +34,6 @@ def send_payment_mail(member, amount, mobilepay_comment):
 
 
 data_sent = {}
-
-
-class fakefile:
-    data = ""
-
-    def write(self, data):
-        self.data += data
-
-
-# little function to make sure the csv data always has the same format
-def rows_to_csv(rows) -> str:
-    file = fakefile()
-    # Converting elements in rows to strings to ensure it can be written to the file object
-    csv.writer(file).writerows([[str(item) for item in row] for row in rows])
-    return file.data
 
 
 def send_userdata_mail(member):
